CLI tool cleanup (#239)
* Renamed module for CLI tool helpers
* Added helper to log and raise errors in a CLI tool
* Changed code to use pathlib
